requirements.txt
python-dotenv
sqlalchemy
greenlet>=3.0
pydantic-settings>=2.0,<3
pydantic>=2.7,<3
fastapi==0.115.5
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
boto3==1.34.156
alembic==1.13.1
asyncmy==0.2.9
pymysql==1.1.0
google-cloud-speech==2.34.0
librosa==0.10.1
numpy==1.26.4
scikit-learn==1.3.2
pydub==0.25.1
torch>=2.2.1
torchaudio>=2.2.1
transformers==4.57.3
peft==0.18.0
sentence-transformers==2.7.0
httpx==0.27.0
chromadb==0.5.11
openai==1.10.0
passlib==1.7.4
python-jose[cryptography]==3.5.0
bcrypt==4.3.0
email-validator==2.3.0